achieve reply function

+++ src/main/java/com/example/demo/controller/PublishController.java
@@ -30,7 +30,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id") Integer id,
+    public String edit(@PathVariable(name="id") Long id,
                         Model model){
         QuestionDTO questionDTO = questionDTOService.getById(id);
         model.addAttribute("title",questionDTO.getTitle());
@@ -45,7 +45,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
@@ -75,7 +75,7 @@ public class PublishController {
         question.setTag(tag);
         question.setDescription(description);
         question.setTitle(title);
-        question.setCreator(Integer.valueOf(user.getId()));
+        question.setCreator(user.getId());
         question.setId(id);
         questionDTOService.createOrUpdate(question);
         return "redirect:/";
